Correction problème d'id du package domain

+++ src/main/java/domain/Person.java
@@ -2,6 +2,8 @@ package domain;
 
 import java.util.Collection;
 
+import javax.persistence.CascadeType;
+import javax.persistence.Column;
 import javax.persistence.Entity;
 import javax.persistence.GeneratedValue;
 import javax.persistence.Id;
@@ -33,7 +35,7 @@ public class Person {
 
 	@Id
 	@GeneratedValue
-	public long getIdPerson() {
+	public int getIdPerson() {
 		return idPerson;
 	}
